Updated naming for distinct names
